# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "EdgeCraft Swagger API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/clouds/{cloudId}/clusters/{clusterId}/nodesets": {
            "get": {
                "description": "클러스터의 NodeSet 리스트 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-NodeSet"
                ],
                "summary": "GetNodeSetList",
                "operationId": "GetNodeSetList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "클러스터에 NodeSet 추가 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-NodeSet"
                ],
                "summary": "SetNodeSet",
                "operationId": "SetNodeSet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "NodeSet Info",
                        "name": "NodeSet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NodeSetInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/clouds/{cloudId}/clusters/{clusterId}/nodesets/{nodeSetId}": {
            "get": {
                "description": "클러스터의 NodeSet 상세 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-NodeSet"
                ],
                "summary": "GetNodeSet",
                "operationId": "GetNodeSet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NodeSet ID",
                        "name": "nodeSetId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "delete": {
                "description": "클러스터의 NodeSet 삭제 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-NodeSet"
                ],
                "summary": "DeleteNodeSet",
                "operationId": "DeleteNodeSet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NodeSet ID",
                        "name": "nodeSetId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/clouds/{cloudId}/clusters/{clusterId}/nodesets/{nodeSetId}/{nodeCount}": {
            "get": {
                "description": "클러스터의 NodeSet에 NodeCount 갱신 (Openstack)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openstack-Cluster-NodeSet"
                ],
                "summary": "UpdateNodeCount",
                "operationId": "UpdateNodeCount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cloud ID",
                        "name": "cloudId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NodeSet ID",
                        "name": "nodeSetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NodeCount",
                        "name": "nodeCount",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.NodeSetInfo": {
            "type": "object",
            "properties": {
                "flavor": {
                    "type": "string",
                    "example": "m1.medium"
                },
                "labels": {
                    "type": "array",
                    "items": {}
                },
                "name": {
                    "type": "string",
                    "example": ""
                },
                "node_count": {
                    "type": "integer",
                    "example": 1
                },
                "nodeset_uid": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "response.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "isError": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}